mobile reader: extract useReaderBook
Lifts book-meta resolve effect (editionId/title/chapters/totalWordCount/
bookmarks load + book_opened analytics) out of the screen.
